+++ plugins/source/azure/resources/services/compute/virtual_machine_scale_set_vm_network_interfaces_test.go
@@ -1,0 +1,34 @@
+package compute
+
+import (
+	"encoding/json"
+	"net/http"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/network/armnetwork/v2"
+	"github.com/cloudquery/plugin-sdk/faker"
+	"github.com/gorilla/mux"
+)
+
+func createVirtualMachineScaleSetsNetworkInterfaces(router *mux.Router) error {
+	var item armnetwork.InterfacesClientListVirtualMachineScaleSetNetworkInterfacesResponse
+	if err := faker.FakeObject(&item); err != nil {
+		return err
+	}
+
+	emptyStr := ""
+	item.NextLink = &emptyStr
+
+	router.HandleFunc("/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/microsoft.Compute/virtualMachineScaleSets/{virtualMachineScaleSetName}/networkInterfaces", func(w http.ResponseWriter, r *http.Request) {
+		b, err := json.Marshal(&item)
+		if err != nil {
+			http.Error(w, "unable to marshal request: "+err.Error(), http.StatusBadRequest)
+			return
+		}
+		if _, err := w.Write(b); err != nil {
+			http.Error(w, "failed to write", http.StatusBadRequest)
+			return
+		}
+	})
+
+	return nil
+}